+++ 1-Authentication/0-sign-in-vanillajs/App/public/redirect.html
