+++ 1-Authentication/0-sign-in-vanillajs/App/server.js
@@ -1,0 +1,35 @@
+const express = require('express');
+const morgan = require('morgan');
+const path = require('path');
+
+const DEFAULT_PORT = process.env.PORT || 3000;
+
+// initialize express.
+const app = express();
+
+// Configure morgan module to log all requests.
+app.use(morgan('dev'));
+
+// Setup app folders.
+app.use(express.static('public'));
+
+// Set up a route for signout.html
+app.get('/signout', (req, res) => {
+    res.sendFile(path.join(__dirname + '/public/signout.html'));
+});
+
+// set up a route for redirect.html
+app.get('/redirect', (req, res) => {
+    res.sendFile(path.join(__dirname + '/public/redirect.html'));
+});
+
+// Set up a route for index.html
+app.get('/', (req, res) => {
+    res.sendFile(path.join(__dirname + '/index.html'));
+});
+
+app.listen(DEFAULT_PORT, () => {
+    console.log(`Sample app listening on port ${DEFAULT_PORT}!`);
+});
+
+module.exports = app;
