+++ 1-Authentication/0-sign-in-vanillajs/AppCreationScripts/AppCreationScripts.md
@@ -4,13 +4,6 @@
 
 ### Quick summary
 
-1. On Windows, run PowerShell as **Administrator** and navigate to the root of the cloned directory
-1. In PowerShell run:
-
-   ```PowerShell
-   Set-ExecutionPolicy -ExecutionPolicy RemoteSigned -Scope Process -Force
-   ```
-
 1. Run the script to create your Azure AD application and configure the code of the sample application accordingly.
 
    ```PowerShell
@@ -59,13 +52,8 @@ The `Configure.ps1` will stop if it tries to create an Azure AD application whic
 
 ### Pre-requisites
 
+1. PowerShell 7 or later (see: [installing PowerShell](https://learn.microsoft.com/en-us/powershell/scripting/install/installing-powershell))
 1. Open PowerShell (On Windows, press  `Windows-R` and type `PowerShell` in the search window)
-1. Navigate to the root directory of the project.
-1. Until you change it, the default [Execution Policy](https:/go.microsoft.com/fwlink/?LinkID=135170) for scripts is usually `Restricted`. In order to run the PowerShell script you need to set the Execution Policy to `RemoteSigned`. You can set this just for the current PowerShell process by running the command:
-
-    ```PowerShell
-    Set-ExecutionPolicy -ExecutionPolicy RemoteSigned -Scope Process
-    ```
 
 ### (Optionally) install Microsoft.Graph.Applications PowerShell modules
 
@@ -73,14 +61,14 @@ The scripts install the required PowerShell module (Microsoft.Graph.Applications
 
 1. If you have never done it already, in the PowerShell window, install the Microsoft.Graph.Applications PowerShell modules. For this:
 
-   1. Open PowerShell as admin (On Windows, Search Powershell in the search bar, right click on it and select **Run as administrator**).
+   1. Open PowerShell
    2. Type:
 
       ```PowerShell
       Install-Module Microsoft.Graph.Applications
       ```
 
-      or if you cannot be administrator on your machine, run:
+      or if you want the modules to be installed for the current user only, run:
 
       ```PowerShell
       Install-Module Microsoft.Graph.Applications -Scope CurrentUser
