+++ 1-Authentication/0-sign-in-vanillajs/ReadmeFiles/topology.png
