--- 1-Authentication/5-sign-in-javascript/App/index.html
@@ -1,68 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0, shrink-to-fit=no">
-  <title>Microsoft identity platform</title>
-  <link rel="SHORTCUT ICON" href="./favicon.svg" type="image/x-icon">
-  <link rel="stylesheet" href="./styles.css">
-
-
-  <!-- msal.min.js can be used in the place of msal.js; included msal.js to make debug easy -->
-  <script id="load-msal" src="https://alcdn.msauth.net/browser/2.31.0/js/msal-browser.js"
-    integrity="sha384-BO4qQ2RTxj2akCJc7t6IdU9aRg6do4LGIkVVa01Hm33jxM+v2G+4q+vZjmOCywYq"
-    crossorigin="anonymous"></script>
-
-  <!-- adding Bootstrap 5 for UI components  -->
-  <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.2.2/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-Zenh87qX5JnK2Jl0vWa8Ck2rdkQ2Bzep5IDxbcnCeuOxjzrPF/et3URy9Bv1WTRi" crossorigin="anonymous">
-</head>
-
-<body>
-  <nav class="navbar navbar-expand-lg navbar-dark bg-primary navbarStyle">
-    <a class="navbar-brand" href="/">Microsoft identity platform</a>
-    <div class="collapse navbar-collapse justify-content-end">
-      <button type="button" id="signIn" class="btn btn-secondary" onclick="signIn()">Sign-in</button>
-      <button type="button" id="signOut" class="btn btn-success d-none" onclick="signOut()">Sign-out</button>
-    </div>
-  </nav>
-  <br>
-  <h5 id="title-div" class="card-header text-center">Vanilla JavaScript single-page application secured with MSAL.js
-  </h5>
-  <h5 id="welcome-div" class="card-header text-center d-none"></h5>
-  <br>
-
-  <div class="table-responsive-ms" id="table">
-    <table id="table-div" class="table table-striped d-none">
-        <thead id="table-head-div">
-        <tr>
-            <th>Claim Type</th>
-            <th>Value</th>
-            <th>Description</th>
-        </tr>
-        </thead>
-        <tbody id="table-body-div">
-        </tbody>
-    </table>
-  </div>
-
-  <!-- importing bootstrap.js and supporting js libraries -->
-  <script src="https://code.jquery.com/jquery-3.3.1.slim.min.js" 
-  integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo" crossorigin="anonymous">
-  </script>
-  <script src="https://cdn.jsdelivr.net/npm/@popperjs/core@2.11.6/dist/umd/popper.min.js" integrity="sha384-oBqDVmMz9ATKxIep9tiCxS/Z9fNfEXiDAYTujMAeBAsjFuCZSmKbSSUnQlmh/jp3" crossorigin="anonymous"></script>
-
-  <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.2.2/dist/js/bootstrap.bundle.min.js" integrity="sha384-OERcA2EqjJCMA+/3y+gxIOqMEjwtxJY7qPCqsdltbNJuaOe923+mo//f6V8Qbsw3" crossorigin="anonymous"></script>
-
-
-  <!-- importing app scripts (load order is important) -->
-  <script type="text/javascript" src="./authConfig.js"></script>
-  <script type="text/javascript" src="./ui.js"></script>
-  <script type="text/javascript" src="./utils/claimUtils.js"></script>
-
-  <!-- <script type="text/javascript" src="./authRedirect.js"></script> -->
-  <!-- uncomment the above line and comment the line below if you would like to use the redirect flow -->
-  <script type="text/javascript" src="./authPopup.js"></script>
-</body>
-
-</html>
